# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Eggpoint — B2B egg supplier directory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      summary: List verified farms
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of farms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    post:
      operationId: createFarm
      tags: [farms]
      summary: Register as a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmInput"
      responses:
        "201":
          description: Farm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /farms/{farmCode}:
    get:
      operationId: getFarm
      tags: [farms]
      summary: Get farm profile
      parameters:
        - name: farmCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Farm profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "404":
          description: Farm not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFarm
      tags: [farms]
      summary: Update farm profile
      parameters:
        - name: farmCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmUpdate"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Farm not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: Get inquiries for current seller's farm
      responses:
        "200":
          description: Inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Send an inquiry to a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Inquiry sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inquiries/{id}:
    patch:
      operationId: updateInquiry
      tags: [inquiries]
      summary: Update inquiry status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryUpdate"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Farm:
      type: object
      required: [id, farmCode, farmName, state, lga, verified, subscriptionTier, createdAt]
      properties:
        id:
          type: integer
        farmCode:
          type: string
        ownerId:
          type: integer
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        farmName:
          type: string
        state:
          type: string
        lga:
          type: string
        description:
          type: ["string", "null"]
        verified:
          type: boolean
        subscriptionTier:
          type: string
          enum: [FREE, FEATURED]
        activeBatchCount:
          type: integer
        totalCratesAvailable:
          type: integer
        createdAt:
          type: string
    FarmInput:
      type: object
      required: [farmName, state, lga]
      properties:
        farmName:
          type: string
          minLength: 2
        state:
          type: string
        lga:
          type: string
        description:
          type: string
    FarmUpdate:
      type: object
      properties:
        farmName:
          type: string
          minLength: 2
        lga:
          type: string
        description:
          type: string
...
    Inquiry:
      type: object
      required: [id, buyerName, buyerPhone, farmCode, batchCode, quantityCrates, message, status, createdAt]
      properties:
        id:
          type: integer
        buyerName:
          type: string
        buyerPhone:
          type: string
        farmCode:
          type: string
        batchCode:
          type: string
        quantityCrates:
          type: integer
        message:
          type: string
        status:
          type: string
          enum: [PENDING, RESPONDED, CLOSED]
        createdAt:
          type: string
    InquiryInput:
      type: object
      required: [buyerName, buyerPhone, farmCode, batchCode, quantityCrates, message]
      properties:
        buyerName:
          type: string
          minLength: 2
        buyerPhone:
          type: string
          minLength: 7
        farmCode:
          type: string
        batchCode:
          type: string
        quantityCrates:
          type: integer
          minimum: 1
        message:
          type: string
          minLength: 10
    InquiryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, RESPONDED, CLOSED]
...